Factor out components and hook and organize.
